Allow Stacked Unary Operators, Change Exponent Precedence
Unary negation operator is now part of the AST and generated code. Changed code to support multiple unary operators stacked on top of each other, like "x = ---4", evaluating to "x = -4."

Exponents now group on the right, so 2^3^4 is 2^(3^4) instead of 2^(3^4), which seems more conventional.

Use the name 'primary' instead of 'basicFactor.' It is a clearer word that avoids some concepts in 'factor' that don't apply.