Changes up to and Including Block Comments Exercise
common.lua's endToken is now a grammar rule instead of a pattern. New endTokenPattern in common.lua for use in the grammar.

Support block comments in the end token pattern.

Add T function to tokenize literals.

Grammar is now self-contained and not modified after creation.

Test files for identifiers and numerals now use grammars with endToken set as a symbol, since endToken is now a grammar symbol.