Keyword Exclude Rules
This includes changes up to and including the exercise "Alternative for Reserved Words."

Identifiers are rejected if they are identical to keywords.

Keywords are created in the grammar with KW(), which also adds them to the list that is checked in the identifiers.

Match identifiers after things that start with keywords.