Fixed Function Scope Bug
Attempt to generate less redundant code resulted in some issues.

Be sure that scope goes up when writing functions and down afterward.

The only thing we want to avoid is popping if we're writing a function, since function returns handle that.
But we do have to save the parameter count since we change the values after leaving the function.