Ternary Operator Support
C-style ?: operator.

A ternary expression has the lowest precedence. It is a logic expression, a question mark, a true expression, a colon, and a false expression, or a logic expression to move up in  the precedence graph.

Four basic test cases for the new operator. True/false evalating correctly, and the two kinds of type checking errors.
Four basic test cases for the new operator. True/false evaluating correctly, and the two kinds of type checking errors.

Require branch types to match, and for the condition to be a boolean.

Logic is similar to if/else, but simpler because only expressions are handled.

Corrected AST and code versions for everything but GraphViz.