Array Function Assignment and Calling
Array element types can be function types.

Functions can be called from targets.

Type checker assigned inferred type to the type in the AST, since it's needed to be able to call functions if the type is inferred.

Write targets are now called 'targets,' which I think is clearer.

Documentation updated to reflect new behavior.