Fix Infinite Loop in Input Program
The stack was being put into an inconsistent state because the interpreter was preserving the top element of the stack as if it was a return value, even though the function returned nothing so there was actually nothing there.
The end result is that it was wiping out what used to be on the top of the stack.

Added a new 'exit' instruction that is coded for functions that have no return values.

The bug was introduced in e4260194caaa3f7e1d2f62f65846fd2c775121c2, "Additional Elegance," as it added support for functions with no return values to not push values onto the stack.