More Traditional Syntax
No need for `call`, `return` expression is optional, assignment operator is `=`.

`return ` is excluded from the beginning of identifiers, and `if ` and `while ` are excluded from the beginning of identifiers that have function types. Note the spaces. `returnSomething` is OK, and `whileFunction` is OK for an identifier used for a function.